requirements.txt
Flask==1.1.2
Flask-Compress==1.5.0
flask-ngrok==0.0.25
Flask-Session==0.3.2
Flask-SocketIO==5.0.1
gunicorn==20.0.4
itsdangerous==1.1.0
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.2
numpydoc==0.9.2
pandas==1.1.3
pandas-datareader==0.6.0
requests==2.23.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==0.9.1
scikit-learn==0.23.1
scikit-multilearn==0.2.0
Werkzeug==1.0.1